Move subprocess launcher into process library.
